web-apps/svg-dropper/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "lxml>=6.0.0",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
    "svgpathtools>=1.7.1",
    "uvicorn>=0.35.0",
    "vpype-gcode>=0.13.0",
]
